+++ lib/src/widgets/picture_picker.dart
@@ -1,7 +1,9 @@
+import 'dart:ui' as ui;
+
 import 'package:flutter/material.dart';
 import 'package:camera/camera.dart';
+import 'package:picture_cropper/src/controllers/picture_cropper_controller.dart';
 
-import '../controllers/picture_cropper_controller.dart';
 import 'crop/camera_crop_guideline.dart';
 
 /// [PicturePicker] is a widget used for taking photos and picking images from the gallery.
@@ -29,8 +31,12 @@ class PicturePicker extends StatefulWidget {
 }
 
 class _PicturePickerState extends State<PicturePicker> {
+  final GlobalKey _renderBoxKey = GlobalKey();
+  Size _renderBoxSize = ui.Size(0, 0);
+
   @override
   void initState() {
+    WidgetsBinding.instance.addPostFrameCallback(_getStackSize);
     widget.controller.addListener(_initializeCamera);
     super.initState();
   }
@@ -45,16 +51,25 @@ class _PicturePickerState extends State<PicturePicker> {
     setState(() {});
   }
 
+  void _getStackSize(_) {
+    final RenderBox renderBox =
+        _renderBoxKey.currentContext?.findRenderObject() as RenderBox;
+    _renderBoxSize = renderBox.size;
+    widget.controller.setRenderBoxSize(_renderBoxSize);
+    setState(() {});
+  }
+
   @override
   Widget build(BuildContext context) {
     return FutureBuilder<void>(
+        key: _renderBoxKey,
         future: widget.controller.initializeControllerFuture,
         builder: (context, snapshot) {
           return Stack(
             children: [
               SizedBox(
-                width: MediaQuery.of(context).size.width,
-                height: MediaQuery.of(context).size.height,
+                width: _renderBoxSize.width,
+                height: _renderBoxSize.height,
                 child: snapshot.connectionState == ConnectionState.done &&
                         widget.controller.cameraController != null &&
                         widget.controller.cameraController!.value.isInitialized